requirements.txt
requests==2.28.1
python-dotenv==1.0.0
tenacity==8.2.3
openai==1.4.0
PyPDF2